refactor: complete delegated skill tool ownership
